lib/vfscore: Use `unsigned long` for `fcntl`'s third argument's type
Although `fcntl` is most frequently used in conjunction with `int`s
as its third argument, the use case involving file locking with
operations such as `F_SETLK` or `F_GETLK` may require that the
third argument is a pointer.

Therefore, allow for correct addresses above 2G to be passed through
`fcntl` by changing the type of the third argument to `unsigned long`.

Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1213

+++ lib/vfscore/main.c
@@ -1892,7 +1892,7 @@ UK_TRACEPOINT(trace_vfs_fcntl, "%p %d %#x", void *, int, int);
 UK_TRACEPOINT(trace_vfs_fcntl_ret, "\"%s\"", int);
 UK_TRACEPOINT(trace_vfs_fcntl_err, "%d", int);
 
-int vfscore_fcntl(struct vfscore_file *fp, unsigned int cmd, int arg)
+int vfscore_fcntl(struct vfscore_file *fp, unsigned int cmd, unsigned long arg)
 {
 	int ret = 0, error = 0;
 	int tmp, oldf;
